Implement query manager
Implement a simple (!) query manager, allowing to interface with the
Onionoo-API to fetch documents. This implementation only uses `simpleHttp` for
querying the API and has absolutely zero support for any of the optimizations
suggested in https://onionoo.torproject.org/protocol.html#general
All queries will return `IO (Either String a)`, with `a` being the document in
request.